requirements.txt
Django==1.6.3
Fabric==1.8.3
Pillow==2.4.0
South==0.8.4
boto==2.28.0
django-oauth2-provider==0.2.6.1
django-storages==1.1.8
django-widget-tweaks==1.3
djangorestframework==2.3.13
ecdsa==0.11
gunicorn==18.0
mock==1.0.1
model-mommy==1.2
paramiko==1.12.4
psycopg2==2.5.2
pycrypto==2.6.1
pytz==2014.2
requests==2.3.0
shortuuid==0.4.2
six==1.3.0
wsgiref==0.1.2
dj-database-url==0.4.0
whitenoise==2.0.5